@@ -4,24 +4,26 @@
   "license": "MIT",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "^4.12.4",
-    "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^13.3.0",
-    "@testing-library/user-event": "^13.5.0",
-    "@types/jest": "^27.5.2",
-    "@types/node": "^16.11.45",
+    "@emotion/react": "^11.11.4",
+    "@emotion/styled": "^11.11.0",
+    "@mui/material": "^5.15.14",
+    "@testing-library/jest-dom": "^6.4.2",
+    "@testing-library/react": "^14.2.2",
+    "@testing-library/user-event": "^14.5.2",
+    "@types/jest": "^29.5.12",
+    "@types/node": "^20.11.30",
     "@types/react": "^18.0.15",
     "@types/react-dom": "^18.0.6",
-    "axios": "^0.27.2",
-    "concurrently": "^7.2.2",
+    "axios": "^1.6.8",
+    "concurrently": "^8.2.2",
     "cors": "^2.8.5",
-    "d3": "5.16.0",
+    "d3": "^7.9.0",
     "dotenv": "^16.0.1",
     "express": "^4.18.1",
     "express-session": "^1.17.3",
     "fs": "^0.0.1-security",
     "igv": "^2.12.6",
-    "jotai": "^1.7.5",
+    "jotai": "^2.7.1",
     "jquery": "^3.6.0",
     "os": "^0.1.2",
     "path": "^0.12.7",
@@ -29,15 +31,14 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.3.0",
-    "react-scripts": "5.0.1",
     "react-table": "^7.8.0",
     "sequelize": "^6.21.3",
     "sequelize-auto": "^0.8.8",
     "serve": "^14.0.1",
     "sqlite3": "^5.0.9",
-    "styled-components": "^5.3.5",
-    "typescript": "^4.7.4",
-    "web-vitals": "^2.1.4"
+    "styled-components": "^6.1.8",
+    "typescript": "^5.4.3",
+    "web-vitals": "^3.5.2"
   },
   "scripts": {
     "copy-webconfig": "cp asset/webpack.config*.js node_modules/react-scripts/config/",
@@ -66,14 +67,18 @@
     ]
   },
   "devDependencies": {
-    "@types/axios": "^0.14.0",
     "@types/cors": "^2.8.12",
-    "@types/dotenv": "^8.2.0",
     "@types/express": "^4.17.13",
     "@types/express-session": "^1.17.5",
     "@types/react-table": "^7.7.12",
     "@types/styled-components": "^5.1.25",
+    "fork-ts-checker-webpack-plugin": "^6.5.3",
+    "react-scripts": "^5.0.1",
     "ts-node": "^10.9.1",
     "ts-node-dev": "^2.0.0"
+  },
+  "overrides": {
+    "fork-ts-checker-webpack-plugin": "^6.5.3",
+    "typescript": "^5.4.3"
   }
 }
